fix worker partitioning

+++ src/main/scala/parameter/server/algorithms/matrix/factorization/TrainAndEvalServerLogic.scala
@@ -9,17 +9,19 @@ import parameter.server.utils.{Types, Vector}
 
 import scala.collection.mutable
 
-class TrainAndEvalServerLogic(_init: Int => Vector, _update: (Vector, Vector) => Vector) extends ServerLogic[Int, Int, Vector] {
+class TrainAndEvalServerLogic(_init: Int => Vector, _update: (Vector, Vector) => Vector) extends ServerLogic[Long, Int, Vector] {
   val model = new mutable.HashMap[Int, Vector]()
 
   @transient lazy val init: Int => Vector = _init
   @transient lazy val update: (Vector, Vector) => Vector = _update
 
-  override def onPullReceive(pull: Messages.Pull[Int, Int, Vector], out: Collector[Either[Types.ParameterServerOutput, Messages.Message[Int, Int, Vector]]]): Unit = {
+  override def onPullReceive(pull: Messages.Pull[Long, Int, Vector],
+                             out: Collector[Either[Types.ParameterServerOutput, Messages.Message[Long, Int, Vector]]]): Unit = {
     out.collect(Right(PullAnswer(pull.dest, pull.src, model.getOrElseUpdate(pull.dest, init(pull.dest)))))
   }
 
-  override def onPushReceive(push: Messages.Push[Int, Int, Vector], out: Collector[Either[Types.ParameterServerOutput, Messages.Message[Int, Int, Vector]]]): Unit = {
+  override def onPushReceive(push: Messages.Push[Long, Int, Vector],
+                             out: Collector[Either[Types.ParameterServerOutput, Messages.Message[Long, Int, Vector]]]): Unit = {
     val oldParam = model(push.destination)
 
     model.update(push.destination, update(oldParam, push.msg))
